# openapi/mocknetpack.yaml
openapi: 3.0.3
info:
  title: MockNetPack API
  version: 0.13.0
  description: |
    MockNetPack 中央 Mock 服务器 API 契约。
    - AI 接入层：为 Agent 工具与脚本/CLI 提供机器原生通道——
      (1) 长效 API Key（POST/GET /api/v1/auth/keys、DELETE /api/v1/auth/keys/{id}）：创建时明文仅返回一次，
      服务端只存 SHA-256 哈希；`Authorization: Bearer <apiKey>` 与登录会话 token 并存，鉴权语义一致；
      (2) 流量数据体验补强：GET /sessions/{id}/traffic 新增 projection=compact（8 字段瘦身视图 +
      每会话自增 seq 增量游标）、since（seq 增量拉取）、method/scheme 过滤，默认（非 compact）响应形状不变，
      人类 Web UI 体验不回归；(3) 错误响应按「给 AI 的指引」：401 提示配置 API Key、404/409/403 说明原因与下一步。
    - 权限与分享（/，v0.11.0）：MockRule 新增 owner（创建者用户名，服务端从当前会话用户写入，
      创建时设）与 updatedBy（最后修改者，每次 PUT 更新）两个纯服务端字段——Web 全量列表下发展示；
      SDK 增量拉取（sinceVersion）只返回 SDK 消费字段（id/method/path/response/enabled/effective），
      绝不携带 owner/updatedBy 等纯服务端/Web-only 字段。权限矩阵：查看规则对设备可见者全部开放
      （设备隔离  不变）；编辑回包/启停 toggle/删除仅规则 owner 与 admin（PUT / DELETE 非 owner
      且非 admin → 403 forbidden，且 403 不泄露规则存在性——对非 admin 未知 ruleId 同样 403）；
      「Mock 此请求」创建新规则所有开发者可用，创建后自己即 owner。存量规则（owner 为空，v0.11.0
      之前创建）仅 admin 可管理，dev 只读 403。分享持久化：ShareSnapshot 随 data.json 落盘，
      重启不失效，7 天 TTL（expiresAt = 创建时刻 + 7 天）不变——过期链接返回 404，janitor 每小时
      清理过期快照（与  48h 会话流量保留并存）。
    - 容量与留存：会话流量上限契约化（--capture-session-max-entries，默认 20000，
      超限丢最旧 rolling window）；服务端过滤下推（GET /sessions/{id}/traffic 新增
      keyword/statusCode/from/to，过滤先于分页）；会话结束后流量保留 48h（--capture-traffic-retention-hours，
      默认 48，自结束时刻起算）供历史日志查询，之后由 janitor 每小时清理；历史日志查询期间保留期内
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /sessions:
    post:
      tags: [sessions]
      summary: 激活抓包会话（Web「连接」）
      description: |
        同一设备同时仅存在一个抓包会话：若已有激活会话则返回现有会话（幂等）。
        会话激活后，SDK 在下一次心跳（≤ 心跳间隔）感知并开始采集。
      operationId: activateSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivateSessionRequest'
      responses:
        '201':
          description: 新建会话
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureSession'
        '200':
          description: 已有激活会话，返回现有会话
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureSession'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 设备离线（device_offline，心跳超时），不可激活会话
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [sessions]
      summary: 会话列表（按设备查询）
      description: 返回该 (app, did) 的会话列表（v0.10.0：capturing + 保留期内 ended 会话；
        历史 ended 会话期满由 janitor 清理后不再返回）。app、did 均为必填 query 参数（隔离强制）。
      operationId: listSessions
      parameters:
        - $ref: '#/components/parameters/AppQuery'
        - $ref: '#/components/parameters/DidQuery'
      responses:
        '200':
          description: 会话列表
          content:
            application/json:
              schema:
                type: object
                required: [sessions, total]
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/CaptureSession'
                  total:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
  /sessions/{id}:
    get:
      tags: [sessions]
      summary: 会话概要（起止时间/请求数/状态/保留截止）
      description: |
        （v0.10.0）：capturing 会话返回实时状态；ended 会话在保留期（retainUntil 之前）内仍返回
        200（status=ended + endedAt + retainUntil，供历史日志展示），期满（即使 janitor 尚未清理）
        返回 404 session_not_found。
      operationId: getSession
      parameters:
        - $ref: '#/components/parameters/SessionIdPath'
      responses:
        '200':
          description: 会话概要（ended 会话保留期内可查）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureSession'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [sessions]
      summary: 强制结束会话（Web「断开」）
      description: |
        结束会话：SDK 停止采集与 Mock；设备回到待命。与"最后一个查看者释放"、"心跳超时"同为会话结束路径。
        （v0.10.0，标记保留而非删除）：会话标记 status=ended + endedAt + retainUntil=now+48h
        （--capture-traffic-retention-hours，默认 48），流量保留在会话内供历史日志查询；
        保留期内 GET /sessions/{id} 与流量列表仍可访问，期满由 janitor 每小时清理释放内存。
        会话结束（任意路径）统一副作用：该设备所有 enabled=true 的 Mock 规则批量置为
        enabled=false 并 Bump 规则版本，规则不删除、仍在规则列表，
        下次建立会话后由用户手动一键重新启用。
      operationId: endSession
      parameters:
        - $ref: '#/components/parameters/SessionIdPath'
      responses:
        '204':
          description: 已结束
        '404':
          $ref: '#/components/responses/NotFound'
  /sessions/{id}/viewers:
    post:
      tags: [sessions]
      summary: 登记/续租查看者（Web 页面）
      description: |
        每个打开该设备页面的浏览器调用一次以登记页面租约；幂等续租（同 viewerId 重复调用刷新 TTL）。
        会话的查看者计数归 0 时（所有页面关闭），会话自动结束（最后一个相关页面关闭才结束）。
        viewerId 由客户端生成（页面级 UUID），页面存活期间周期性续租，关闭时释放。
      operationId: registerViewer
      parameters:
        - $ref: '#/components/parameters/SessionIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterViewerRequest'
      responses:
        '200':
          description: 登记或续租成功，返回租约信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ViewerLease'
        '404':
          $ref: '#/components/responses/NotFound'
  /sessions/{id}/viewers/{viewerId}:
    delete:
      tags: [sessions]
      summary: 释放查看者（页面关闭）
      description: 页面关闭/卸载时调用；释放后服务端重算计数，归 0 则结束会话。
      operationId: releaseViewer
      parameters:
        - $ref: '#/components/parameters/SessionIdPath'
        - $ref: '#/components/parameters/ViewerIdPath'
      responses:
        '204':
          description: 已释放
        '404':
          $ref: '#/components/responses/NotFound'

  # ==========================================================================
  # 3. 流量上传 / 查询（ 实现，本阶段仅契约）
  # ==========================================================================
  /traffic:
    post:
      tags: [traffic]
      summary: 批量上传流量（SDK）
      description: |
        SDK 在抓包会话有效期间批量上传一次/多次请求的完整信息（全量抓包，会话内临时记录）。
        攒批：SDK 按条目数或累计字节阈值触发上传，单批 ≤ 500 条；
        服务端请求体上限 10MB（mockd Admin BodyLimit），SDK 应控制批次大小避免整批被拒。
        部分接受：entries 中字段不合法的条目被丢弃、合法条目正常入库，响应 count = 实际接受数；
        全部条目不合法返回 400。
        临时语义：会话流量按 MaxSessionTrafficEntries
        （--capture-session-max-entries，默认 20000）上限保留，超限丢最旧（rolling window）；
        会话结束（页面关闭 / 心跳超时 / 强制断开）后流量保留 48h（--capture-traffic-retention-hours，
        默认 48，自结束时刻起算）供历史日志查询，期满由 janitor 清理；
        命中 Mock 的请求仍按临时流量上传（mocked=true，其响应即 Mock 回包），随会话保留；
        持久化保留的是规则及其 source 快照（原始值+Mock 值），不是每条命中流量。
        仅接受激活会话内的流量：sessionId 不存在 → 404 session_not_found；会话已结束 → 409
        session_ended（保留期内不可再上传）。
      operationId: uploadTraffic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrafficUploadRequest'
      responses:
        '202':
          description: 已接收（含部分接受；count 为实际接受条目数）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrafficUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 会话已结束（session_ended）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /traffic/{id}:
    get:
      tags: [traffic]
      summary: 请求详情
      description: |
        返回单条请求完整信息。会话结束后流量转入保留区（默认 7 天内仍可按 ID 解析，供创建分享
        链接）；超过保留窗口或不存在的记录返回 404（not_found）。
      operationId: getTraffic
      parameters:
        - $ref: '#/components/parameters/TrafficIdPath'
      responses:
        '200':
          description: 请求完整信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrafficEntry'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [traffic]
      summary: 删除单条流量（后端管理用途；Web 自 v0.10.0 起不再调用——「清空/✕」为纯前端操作）
      description: |
        删除一条会话内临时流量记录，仅作用于日志展示；不影响任何已持久化的 Mock 规则
        （日志与规则互相独立）。（v0.10.0）起标注为后端管理用途：Web 不再调用（前端删除只清
        pageLog），服务端保留该接口供运维清理。条目不存在返回 404 not_found，调用方按幂等容错处理。
      operationId: deleteTraffic
      parameters:
        - $ref: '#/components/parameters/TrafficIdPath'
      responses:
        '204':
          description: 已删除
        '404':
          $ref: '#/components/responses/NotFound'
  /shares:
    post:
      tags: [shares]
      summary: 创建分享链接（快照）
      description: |
        把一条流量（进行中会话内，或结束会话的保留区内）复制为独立只读快照并生成分享链接
        。快照自包含（深度拷贝，切断与原会话/流量的关联）。持久化语义：
        快照随 data.json 落盘、服务重启不失效；有效期 7 天（expiresAt = 创建时刻 + 7 天），过期链接
        返回 404，janitor 每小时清理过期快照（与  48h 会话流量保留并存）。
        需要登录（设备可见者：设备 owner 或 admin）。链接本身公开只读（GET /shares/{id} 免登录）。
      operationId: createShare
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShareRequest'
      responses:
        '201':
          description: 创建成功，返回分享链接（7 天有效）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateShareResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /shares/{shareId}:
    get:
      tags: [shares]
      summary: 读取分享快照（公开只读，免登录）
      description: |
        按分享 ID 返回只读快照。持久化语义：链接跨服务重启有效（快照落盘）；
        过期（创建后 7 天）或未知 ID 一律 404。
      operationId: getShare
      parameters:
        - name: shareId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 分享快照
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareSnapshot'
        '404':
          $ref: '#/components/responses/NotFound'
  /sessions/{id}/traffic:
    get:
      tags: [traffic]
      summary: 会话内流量列表（实时 / 历史日志，服务端过滤下推 + 分页 + 机器通道）
      description: |
        Web 请求流数据源（实时）与历史日志查询共用本接口：
        - 实时：Web 每 2s 轮询本接口实现秒级可见（需求），limit 拉取最新条目并按 id 本地去重。
        - 历史日志：capturing 会话实时查询不变；ended 会话在保留期（retainUntil 之前）内同样返回真实
          流量（供最近 48h 历史日志），期满返回 404 session_not_found。
        - 过滤：keyword（URL 子串，大小写不敏感）/ statusCode（精确）/
          from、to（RFC3339，含边界）可任意组合（AND）；非法参数 400 invalid_field。
        - 分页：先过滤后分页——total 为过滤后集合大小；Web「加载更多」用 offset 累加。
        排序：按 timestamp 升序返回（请求时间线）；Web 展示可自行倒序。
        - 机器通道：projection=compact 返回瘦身视图 CompactTrafficView（8 字段，
          不含 headers/bodies/query 等大字段），并为每条补 seq（每会话自增增量游标，ID 字典序不严格单调，
          故新增独立 seq）；since 按 seq 增量拉取（since=0 全量）；method（大小写不敏感）/
          scheme（URL 前缀 http|https）过滤。默认无 projection 时响应形状与 v0.11.0 完全一致
          （TrafficListResponse 完整字段），人类 Web UI 不回归。
      operationId: listSessionTraffic
      parameters:
        - $ref: '#/components/parameters/SessionIdPath'
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
          description: 返回条数上限
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: 跳过条数（先过滤后分页）
        - name: keyword
          in: query
          schema:
            type: string
          description: URL 子串过滤（大小写不敏感）
        - name: statusCode
          in: query
          schema:
            type: integer
          description: 状态码精确过滤
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: 时间范围起点（RFC3339，含边界）
        - name: to
          in: query
          schema:
            type: string
            format: date-time
          description: 时间范围终点（RFC3339，含边界）
        - name: projection
          in: query
          schema:
            type: string
            enum: [compact]
          description: |
            compact 时返回瘦身视图（CompactTrafficView 数组，含 seq 增量游标），
            供 MCP/CLI 等机器通道低带宽拉取；缺省时返回完整 TrafficEntry（v0.11.0 形状不变）。
        - name: since
          in: query
          schema:
            type: integer
            minimum: 0
          description: |
            增量游标——只返回 seq > since 的条目（跨请求批次单调递增），
            用于 MCP/CLI 轮询增量拉取；since=0（缺省）返回全量。与 offset 组合时先按 since 过滤再分页。
        - name: method
          in: query
          schema:
            type: string
          description: HTTP 方法精确过滤（大小写不敏感，如 GET/post）
        - name: scheme
          in: query
          schema:
            type: string
          description: ：URL scheme 过滤（http|https，按 URL 前缀匹配，忽略大小写）
      responses:
        '200':
          description: 流量列表（默认完整条目；projection=compact 时为瘦身条目；均含过滤后 total）
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    required: [entries, total]
                    properties:
                      entries:
                        type: array
                        items:
                          $ref: '#/components/schemas/TrafficEntry'
                      total:
                        type: integer
                  - $ref: '#/components/schemas/TrafficCompactListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [traffic]
      summary: 清空会话流量（后端管理用途；Web 自 v0.10.0 起不再调用——「清空/✕」为纯前端操作）
      description: |
        清空会话的全部临时流量并把会话请求计数重置为 0（时间线重开，新流量继续累积）。
        仅作用于日志展示，不影响任何已持久化的 Mock 规则。（v0.10.0）起标注为后端管理用途：
        Web 不再调用（前端清空只清 pageLog），服务端保留该接口供运维清理。未知会话返回 404。
      operationId: clearSessionTraffic
      parameters:
        - $ref: '#/components/parameters/SessionIdPath'
      responses:
        '204':
          description: 已清空
        '404':
          $ref: '#/components/responses/NotFound'

  # ==========================================================================
  # 4. Mock 规则（下发与 CRUD，/ 实现，本阶段仅契约）
  # ==========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
      description: |
        两种凭证并存：
        - 会话 token：登录（POST /auth/login）签发、服务端持久化；Web 每次请求携带
          `Authorization: Bearer <token>`。有效期 7 天（服务端配置项）；登出即服务端吊销。
        - 长效 API Key：POST /auth/keys 创建，`Authorization: Bearer <mnpk_...>`；无过期，
          明文仅创建时返回一次，服务端只存 SHA-256 哈希（S3：任何日志不落 Authorization/key 明文）。
        两者权限等价（API Key 归属用户角色），401/403 语义一致。
  parameters:
...
    AppQuery:
      name: app
      in: query
      required: true
      schema:
        type: string
      description: App 标识（隔离强制）
    DidQuery:
      name: did
      in: query
      required: true
      schema:
        type: string
      description: 设备标识（隔离强制）
    SessionIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: 抓包会话 ID（ULID）
    ViewerIdPath:
      name: viewerId
      in: path
      required: true
      schema:
        type: string
      description: 查看者租约 ID（页面级 UUID，客户端生成）
...
    TrafficIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: 流量记录 ID
...
  responses:
    BadRequest:
      description: 请求参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: 服务端错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    SessionStatus:
      type: string
      enum: [capturing, ended]
      description: capturing=抓包中；ended=已结束（ 起仅瞬时状态：会话结束即删除记录，不持久化 ended）
    CaptureSession:
      type: object
      required: [id, app, did, status, startedAt]
      properties:
        id:
          type: string
          description: 会话 ID（ULID）
        app:
          type: string
        did:
          type: string
        status:
          $ref: '#/components/schemas/SessionStatus'
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        requestCount:
          type: integer
          default: 0
          description: 会话内请求数（启用后有真实值）
        viewerCount:
          type: integer
          default: 0
          description: 当前打开该设备页面的查看者数（最后一个释放则会话结束）
        retainUntil:
          type: string
          format: date-time
          nullable: true
          description: 结束会话的保留截止时刻
    ActivateSessionRequest:
      type: object
      required: [app, did]
      properties:
        app:
          type: string
          maxLength: 128
        did:
          type: string
          maxLength: 128
    RegisterViewerRequest:
      type: object
      required: [viewerId]
      properties:
        viewerId:
          type: string
          description: 页面级 UUID（客户端生成；重复调用 = 续租）
        label:
          type: string
          description: 页面标识（如浏览器标签标题，调试用）
    ViewerLease:
      type: object
      required: [viewerId, ttlSeconds, expiresAt]
      properties:
        viewerId:
          type: string
        ttlSeconds:
          type: integer
          default: 120
          description: 租约 TTL；页面应周期性续租（建议 TTL/2 间隔）
        expiresAt:
          type: string
          format: date-time
          description: 租约过期时间（服务端对过期租约兜底清理）

    # ---------------------------------------------------------------- 流量
    TrafficEntry:
      type: object
      required: [timestamp, method, url, durationMs]
      properties:
        id:
          type: string
          description: 流量记录 ID（服务端生成）
        sessionId:
          type: string
          description: 所属会话 ID（服务端写入）
        timestamp:
          type: string
          format: date-time
          description: 请求发起时间（RFC3339，SDK 打点）
        method:
          type: string
          example: POST
        url:
          type: string
          description: 完整请求 URL
        path:
          type: string
          description: URL 路径（用于 Mock 匹配）
        query:
          type: string
          description: Query 字符串（不含 ?，原始顺序；无则空串）
        requestHeaders:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: 请求头（多值）
        requestBody:
          type: string
          description: |
            请求体：按 UTF-8 文本保存，超过 1MB 截断（ 固定上限，不做配置项）；
            无法 UTF-8 解码的二进制内容以 "[binary N bytes]" 占位保存。
        requestBodyBase64:
          type: string
          description: 二进制请求体的 base64（无法 UTF-8 解码时填充），供"Mock 此请求"回放原始字节
        requestBodyDecoded:
          type: string
          description: |
            请求体经 App 注入的协议解码器解出的可读 UTF-8 文本。
            仅作 Web 展示；存在时 Web 优先展示它，无则回退 requestBody 的 "[binary N bytes]" 占位。
            不参与回放；解码失败/未注入解码器时缺省。
        statusCode:
          type: integer
          nullable: true
          description: 响应状态码（请求失败时为空）
        responseHeaders:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          nullable: true
          description: 响应头（请求失败时为空）
        responseBody:
          type: string
          nullable: true
          description: |
            响应体：按 UTF-8 文本保存，超过 1MB 截断（ 固定上限，不做配置项）；
            无法 UTF-8 解码的二进制内容以 "[binary N bytes]" 占位保存；请求失败时为空。
        responseBodyBase64:
          type: string
          description: 二进制响应体的 base64（无法 UTF-8 解码时填充），供"Mock 此请求"回放原始字节；请求失败时为空
        responseBodyDecoded:
          type: string
          description: |
            响应体经 App 注入的协议解码器解出的可读 UTF-8 文本。
            仅作 Web 展示；存在时 Web 优先展示它，无则回退 responseBody 的 "[binary N bytes]" 占位。
            不参与回放——回放永远走 responseBodyBase64 原始字节；解码失败/未注入解码器时缺省。
        error:
          type: string
          nullable: true
          description: 错误信息（请求失败时；成功时为空）
        durationMs:
          type: integer
          description: 请求耗时（毫秒，请求发起至收到完成回调）
        mocked:
          type: boolean
          default: false
          description: 是否命中 Mock（ 恒为 false； 起标记，Web 请求流标注"Mock 命中"）
        matchedRuleId:
          type: string
          description: >-
            命中的 Mock 规则 ID（v0.13.0， P1-1）：仅 mocked=true 时非空，
            用于确认该请求究竟命中哪一条规则；真实流量为空/缺省。
        seq:
          type: integer
          description: |
            每会话自增的到达序号：服务端按上传到达顺序从 1 递增，
            供增量拉取（since）使用——ID 为 ULID，同毫秒内字典序不严格单调，故新增独立 seq。
            仅服务端与查询方（MCP/CLI/Web）消费，SDK 不感知、不消费。

    # -------------------------------------------------  机器通道（v0.12.0）
...
    TrafficUploadRequest:
      type: object
      required: [app, did, sessionId, entries]
      properties:
        app:
          type: string
          maxLength: 128
        did:
          type: string
          maxLength: 128
        sessionId:
          type: string
          description: 目标会话 ID（上传仅接受激活会话内流量）
        entries:
          type: array
          maxItems: 500
          items:
            $ref: '#/components/schemas/TrafficEntry'
          description: |
            一次批量上传的流量条目。SDK 攒批：按条目数或累计字节阈值触发上传，
            单批 ≤ 500 条且控制总字节，避免超过服务端 10MB 请求体上限。
    TrafficUploadResponse:
      type: object
      required: [accepted, count]
      properties:
        accepted:
          type: boolean
          description: 是否已接收（202 隐含接收， 恒为 true；false 仅作扩展预留）
        count:
          type: integer
          description: 实际接受的条目数（部分接受：合法条目数；全部条目不合法时服务端返回 400 而非 202）

    # ------------------------------------------------------------ Mock 规则
...
    CreateShareRequest:
      type: object
      required: [trafficId]
      properties:
        trafficId:
          type: string
          description: 被分享流量条目的 ID（进行中会话内，或结束会话 48h 保留区内）
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: 机器可读错误码（如 device_not_registered）
        message:
          type: string
          description: 人类可读错误信息